some refactoring -> lca_utils
