+++ sourmash_lib/lca/command_classify.py
@@ -12,7 +12,6 @@ import csv
 from collections import defaultdict, Counter
 import itertools
 import pprint
-import json
 try:
     from itertools import zip_longest
 except ImportError:
@@ -20,213 +19,12 @@ except ImportError:
 import gzip
 
 import sourmash_lib
+from . import lca_utils
 from ..logging import notify, error
 from .. import sourmash_args
+from .lca_utils import set_debug, debug
 
 DEFAULT_THRESHOLD=5                  # how many counts of a taxid at min
-
-taxlist = ['superkingdom', 'phylum', 'class', 'order', 'family', 'genus',
-           'species']
-
-_print_debug = False
-def debug(*args):
-    if _print_debug:
-        print(*args)
-
-
-def build_tree(assignments, initial=None):
-    """
-    Builds a tree of dictionaries from lists of (rank, name) tuples
-    in 'assignments'.  This tree can then be used to find least common
-    ancestor agreements/confusion.
-    """
-    if initial is None:
-        tree = {}
-    else:
-        tree = initial
-
-    for assignment in assignments:
-        node = tree
-
-        for rank, name in assignment:
-            if name:
-                child = node.get((rank, name), {})
-                node[(rank, name)] = child
-
-                # shift -> down in tree
-                node = child
-
-    return tree
-
-
-def test_build_tree():
-    tree = build_tree([[('rank1', 'name1'), ('rank2', 'name2')]])
-    assert tree == { ('rank1', 'name1'): { ('rank2', 'name2') : {}} }
-
-
-def test_build_tree_2():
-    tree = build_tree([[('rank1', 'name1'), ('rank2', 'name2a')],
-                       [('rank1', 'name1'), ('rank2', 'name2b')],
-                      ])
-
-    assert tree == { ('rank1', 'name1'): { ('rank2', 'name2a') : {},
-                                           ('rank2', 'name2b') : {}} }
-
-
-def test_build_tree_3():                  # empty 'rank2' name
-    tree = build_tree([[('rank1', 'name1'), ('rank2', '')]])
-    assert tree == { ('rank1', 'name1'): {} }
-
-
-def test_build_tree_4():
-    tree = build_tree([[('rank1', 'name1'), ('rank2', 'name2a')],
-                      ])
-
-    tree = build_tree([[('rank1', 'name1'), ('rank2', 'name2b')],
-                      ], tree)
-
-    assert tree == { ('rank1', 'name1'): { ('rank2', 'name2a') : {},
-                                           ('rank2', 'name2b') : {}} }
-
-
-def find_lca(tree):
-    """
-    Given a tree produced by 'find_tree', find the first node with multiple
-    children, OR the only leaf in the tree.  Return ((rank, name), reason),
-    where 'reason' is the number of children of the returned node, i.e.e
-    0 if it's a leaf and > 1 if it's an internal node.
-    """
-
-    node = tree
-    cur = ('root', 'root')
-    while 1:
-        if len(node) == 1:                # descend to only child
-            cur = next(iter(node.keys()))
-            node = node[cur]
-        elif len(node) == 0:              # at leaf; end
-            return cur, 0
-        else:                             # len(node) > 1 => confusion!!
-            return cur, len(node)
-
-
-def test_find_lca():
-    tree = build_tree([[('rank1', 'name1'), ('rank2', 'name2')]])
-    lca = find_lca(tree)
-
-    assert lca == (('rank2', 'name2'), 0)
-
-
-def test_find_lca_2():
-    tree = build_tree([[('rank1', 'name1'), ('rank2', 'name2a')],
-                       [('rank1', 'name1'), ('rank2', 'name2b')],
-                      ])
-    lca = find_lca(tree)
-
-    assert lca == (('rank1', 'name1'), 2)
-
-
-def build_reverse_tree(assignments, initial=None):
-    """
-    Builds a child -> parent dictionary (a reverse DAG) from lists of
-    (rank, name) tuples in 'assignments'.
-    """
-    if initial is None:
-        parents = {}
-    else:
-        parents = initial
-
-    for assignment in assignments:
-        last_node = ('root', 'root')
-        for rank, name in assignment:
-            if name:
-                parents[(rank, name)] = last_node
-                last_node = (rank, name)
-
-    return parents
-
-
-def test_build_reverse_tree():
-    parents = build_reverse_tree([[('rank1', 'name1'), ('rank2', 'name2')]])
-
-    print(parents)
-    assert parents == { ('rank2', 'name2'): ('rank1', 'name1'),
-                        ('rank1', 'name1'): ('root', 'root') }
-
-
-def test_build_reverse_tree_2():
-    parents = build_reverse_tree([[('rank1', 'name1'), ('rank2', 'name2a')],
-                                 [('rank1', 'name1'), ('rank2', 'name2b')],
-                                 ])
-
-    assert parents == { ('rank2', 'name2a'): ('rank1', 'name1'),
-                        ('rank2', 'name2b'): ('rank1', 'name1'),
-                        ('rank1', 'name1'): ('root', 'root') }
-
-
-def test_build_reverse_tree_3():
-    parents = build_reverse_tree([[('rank1', 'name1'), ('rank2', 'name2a')],
-                                 ])
-    parents = build_reverse_tree([[('rank1', 'name1'), ('rank2', 'name2b')],
-                                 ], parents)
-
-    assert parents == { ('rank2', 'name2a'): ('rank1', 'name1'),
-                        ('rank2', 'name2b'): ('rank1', 'name1'),
-                        ('rank1', 'name1'): ('root', 'root') }
-
-
-def test_build_reverse_tree_4():          # empty 'rank2' name
-    parents = build_reverse_tree([[('rank1', 'name1'), ('rank2', '')]])
-
-    print(parents)
-    assert parents == { ('rank1', 'name1'): ('root', 'root') }
-
-
-class LCA_Database(object):
-    def __init__(self):
-        self.lineage_dict = None
-        self.hashval_to_lineage_id = None
-        self.ksize = None
-        self.scaled = None
-        self.signatures_to_lineage = None
-
-    def load(self, db_name):
-        xopen = open
-        if db_name.endswith('.gz'):
-            xopen = gzip.open
-
-        with xopen(db_name, 'rt') as fp:
-            load_d = json.load(fp)
-            version = load_d['version']
-            assert version == '1.0'
-
-            type = load_d['type']
-            assert type == 'sourmash_lca'
-
-            ksize = load_d['ksize']
-            scaled = load_d['scaled']
-
-            lineage_dict_2 = load_d['lineages']
-            lineage_dict = {}
-            for k, v in lineage_dict_2.items():
-                vv = []
-                for rank in taxlist:
-                    name = v.get(rank, '')
-                    vv.append((rank, name))
-
-                lineage_dict[int(k)] = tuple(vv)
-
-            hashval_to_lineage_id_2 = load_d['hashval_assignments']
-            hashval_to_lineage_id = {}
-            for k, v in hashval_to_lineage_id_2.items():
-                hashval_to_lineage_id[int(k)] = v
-
-            signatures_to_lineage = load_d['signatures_to_lineage']
-
-        self.lineage_dict = lineage_dict
-        self.hashval_to_lineage_id = hashval_to_lineage_id
-        self.ksize = ksize
-        self.scaled = scaled
-        self.signatures_to_lineage = signatures_to_lineage
 
 
 def classify_signature(query_sig, dblist, threshold):
@@ -258,15 +56,15 @@ def classify_signature(query_sig, dblist, threshold):
         # for each list of tuple_info [(rank, name), ...] build
         # a tree that lets us discover least-common-ancestor.
         tuple_info = these_assignments[hashval]
-        tree = build_tree(tuple_info)
+        tree = lca_utils.build_tree(tuple_info)
 
         # also update a tree that we can ascend from leaves -> parents
         # for all assignments for all hashvals
-        parents = build_reverse_tree(tuple_info, parents)
+        parents = lca_utils.build_reverse_tree(tuple_info, parents)
 
         # now find either a leaf or the first node with multiple
         # children; that's our least-common-ancestor node.
-        lca, reason = find_lca(tree)
+        lca, reason = lca_utils.find_lca(tree)
         counts[lca] += 1
 
     # ok, we now have the LCAs for each hashval, and their number
@@ -294,7 +92,7 @@ def classify_signature(query_sig, dblist, threshold):
         debug(n, count, xx[1:])
 
         # update tree with this set of assignments
-        build_tree([xx], tree)
+        lca_utils.build_tree([xx], tree)
 
     if n > 1:
         debug('XXX', n)
@@ -304,7 +102,7 @@ def classify_signature(query_sig, dblist, threshold):
         return [('', '')], status
 
     # now find LCA? or whatever.
-    lca, reason = find_lca(tree)
+    lca, reason = lca_utils.find_lca(tree)
     if reason == 0:               # leaf node
         status = 'found'
         debug('END', lca)
@@ -347,8 +145,7 @@ def classify(args):
         sys.exit(-1)
 
     if args.debug:
-        global _print_debug
-        _print_debug = True
+        set_debug(args.debug)
 
     ksize_vals = set()
     scaled_vals = set()
@@ -363,7 +160,7 @@ def classify(args):
         notify('... loading database {}'.format(db_name), end='\r',
               file=sys.stderr)
 
-        lca_db = LCA_Database()
+        lca_db = lca_utils.LCA_Database()
         lca_db.load(db_name)
 
         ksize_vals.add(lca_db.ksize)
@@ -394,7 +191,7 @@ def classify(args):
         csvfp = csv.writer(args.output)
     else:
         notify("outputting classifications to stdout")
-    csvfp.writerow(['ID','status'] + taxlist)
+    csvfp.writerow(['ID','status'] + lca_utils.taxlist)
 
     total_count = 0
     n = 0
@@ -415,7 +212,9 @@ def classify(args):
 
             # output!
             row = [query_sig.name(), status]
-            for taxrank, (rank, name) in zip_longest(taxlist, lineage, fillvalue=('', '')):
+            for taxrank, (rank, name) in zip_longest(lca_utils.taxlist,
+                                                     lineage,
+                                                     fillvalue=('', '')):
                 if rank:
                     assert taxrank == rank
                 row.append(name)
